Create tables in the specified schema to avoid moving the tables afterwards (#21)
Fixes #18. 

* Execute "moveTableToSchema" if table was succesfully created
* Function for getting the connection parameters added
* Function buildConnectionString deleted

This is because the parameters are now obtained through the function
getConnectionParameters

* dbConnectionParam and moveTableToSchema replaced by single function

dbConnectionParam has been deleted because now all parameters are
obtained by a single function which is called in the function
handleTable.

moveTableToSchema has been deleted because tables are now created in
the specified schema. Therefore, there is no need to move the table
after their creation.

Co-authored-by: rdrg109 <>